Rename WDIO adapter filename to match class name
